处理部分反编译代码出现的无用内容，改前后端分离并完成群组部分。

+++ src/main/java/net/convnet/server/core/PacketEncoder.java
@@ -32,7 +32,7 @@ final class PacketEncoder extends MessageToByteEncoder<ResponseReader> {
          LOG.debug("Send packet: [" + out.toString(Constants.CHARSET) + "]");
       }
 
-      Long writeBytes = (Long)ctx.channel().attr(Constants.WRITE_BYTES_KEY).get();
+      Long writeBytes = ctx.channel().attr(Constants.WRITE_BYTES_KEY).get();
       if (writeBytes == null) {
          writeBytes = 0L;
       }
